Refactor(Bug & Feedback): Restructured FloatingBugButton into modular components
